University/requirements.txt
asgiref             
backports.zoneinfo  
certifi             
charset-normalizer  
coreapi             
coreschema          
Django              
django-rest-swagger 
djangorestframework 
idna                
itypes              
Jinja2              
MarkupSafe          
openapi-codec       
pip                 
pkg-resources       
pytz                
requests            
setuptools          
simplejson          
sqlparse            
typing-extensions   
uritemplate         
urllib3             
wheel               